requirements.txt
alembic==1.14.0
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
mysqlclient==2.2.5
packaging==24.1
python-decouple==3.8
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.2
WTForms==3.2.1
